refactor: validation 로직 클래스 위치 exception -> util 하위로 변경
