+++ src/main/java/ui/InputView.java
@@ -1,6 +1,7 @@
 package ui;
 
 import util.Menu;
+import util.ValidationInput;
 
 import java.util.Scanner;
