+++ src/main/java/util/ValidationEquation.java
@@ -1,4 +1,4 @@
-package calculator.exception;
+package util;
 
 import util.OperatorMap;
