+++ src/test/java/calculator/ui/InputViewTest.java
@@ -5,7 +5,7 @@ import exception.NotMenuFormatExcpetion;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import ui.ValidationInput;
+import util.ValidationInput;
 import util.Menu;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
